+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`0.5.0...main`][0.5.0...main].
+For a full diff see [`0.6.0...main`][0.6.0...main].
+
+## [`0.6.0`][0.6.0]
+
+For a full diff see [`0.5.0...0.6.0`][0.5.0...0.6.0].
+
+### Added
+
+- Allowed registering already instantiated entity definition providers ([#702]), by [@localheinz]
 
 ## [`0.5.0`][0.5.0]
 
@@ -171,6 +179,7 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [0.3.2]: https://github.com/ergebnis/factory-bot/releases/tag/0.3.2
 [0.4.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.4.0
 [0.5.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.5.0
+[0.6.0]: https://github.com/ergebnis/factory-bot/releases/tag/0.6.0
 
 [fa9c564...0.1.0]: https://github.com/ergebnis/factory-bot/compare/fa9c564...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/factory-bot/compare/0.1.0...0.2.0
@@ -180,7 +189,8 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [0.3.1...0.3.2]: https://github.com/ergebnis/factory-bot/compare/0.3.1...0.3.2
 [0.3.2...0.4.0]: https://github.com/ergebnis/factory-bot/compare/0.3.2...0.4.0
 [0.4.0...0.5.0]: https://github.com/ergebnis/factory-bot/compare/0.4.0...0.5.0
-[0.5.0...main]: https://github.com/ergebnis/factory-bot/compare/0.5.0...main
+[0.5.0...0.6.0]: https://github.com/ergebnis/factory-bot/compare/0.5.0...0.6.0
+[0.6.0...main]: https://github.com/ergebnis/factory-bot/compare/0.6.0...main
 
 [#1]: https://github.com/ergebnis/factory-bot/pull/1
 [#3]: https://github.com/ergebnis/factory-bot/pull/3
@@ -260,6 +270,7 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [#498]: https://github.com/ergebnis/factory-bot/pull/498
 [#499]: https://github.com/ergebnis/factory-bot/pull/499
 [#682]: https://github.com/ergebnis/factory-bot/pull/682
+[#702]: https://github.com/ergebnis/factory-bot/pull/702
 
 [@abenerd]: https://github.com/abenerd
 [@localheinz]: https://github.com/localheinz